Abstract capability component
