+++ src/components/Capability/index.js
@@ -1,0 +1,1 @@
+export { default } from './Capability';
